==> notebooks/01_load_data.ipynb <==
!pip install polars

import polars as pl
print(pl.__version__)

import numpy as np 
import pandas as pd 
import matplotlib.pyplot as plt

s3_path = "s3://amazon-electronics-dataset/Raw_Dataset/Electronics_5.json"
df = pl.scan_ndjson(s3_path)

df.head().collect()

df.schema

df.describe()

df.count().collect()

df_nulls = df.select([
    pl.col(col).is_null().sum().alias(f"{col}_nulls")
    for col in df.schema.keys()
])

df_nulls.collect()
