==> notebooks/02_preprocess.ipynb <==
!pip install polars

import polars as pl
import pandas as pd
import sys

sys.path.append("../src")
from preprocess import clean_text, flatten_style

#Load raw JSON
BUCKET = "amazon-electronics-dataset"
KEY = "Raw_Dataset/Electronics_5.json"

df_lazy = pl.scan_ndjson(f"s3://{BUCKET}/{KEY}")
df_lazy

#Selecting required columns
df_lazy = df_lazy.select([
    "reviewerID",
    "asin",
    "reviewText",
    "summary",
    "overall",
    "vote",
    "verified",
    "reviewTime",
    "unixReviewTime",
    "style",
    "image",
    "reviewerName"
])

df = df_lazy.collect()
df.head()

#Dropping rows with missing fields
df = (
    df
    .filter(pl.col("asin").is_not_null())
    .filter(pl.col("reviewText").is_not_null())
)

#Rename columns
df = df.rename({
    "reviewerID": "reviewer_id",
    "asin": "product_id",
    "reviewText": "original_review",
    "summary": "user_summary",
    "overall": "rating",
    "vote": "helpful_votes",
    "verified": "is_verified",
    "reviewTime": "review_date",
    "unixReviewTime": "review_timestamp"
})

#Helpful votes
df = df.with_columns([
    pl.col("helpful_votes")
      .str.replace(",", "")
      .cast(pl.Int64, strict=False)
      .fill_null(0)
])

#Fill null summaries
df = df.with_columns([
    pl.col("user_summary")
      .fill_null("No summary provided")
])

#Flatten style metadata
# Flatten style metadata
df = df.with_columns([
    pl.col("style")
      .map_elements(lambda x: flatten_style(x))
      .alias("product_style")
])

#Clean review text
df = df.with_columns([
    pl.col("original_review")
      .map_elements(lambda x: clean_text(x))
      .alias("cleaned_review")
])

#Convert Dates & Timestamp
df = df.with_columns([
    pl.col("review_date")
      .str.to_date(format="%m %d, %Y", strict=False)
      .alias("review_date"),

    pl.col("review_timestamp").cast(pl.Int64)
])


#Drop useless columns
df = df.drop(["image", "reviewerName", "style"])

#Filtering short reviews
df = df.filter(pl.col("cleaned_review").str.len_chars() > 20)

#Filter products with fewer than 5 reviews
counts = (
    df
    .group_by("product_id")
    .count()
    .rename({"count": "review_count"})
)

df = df.join(counts, on="product_id")
df = df.filter(pl.col("review_count") >= 5)

#Save as CSV 
output_csv = "electronics_clean.csv"

df.to_pandas().to_csv(output_csv, index=False)

# Upload to S3
!aws s3 cp electronics_clean.csv s3://amazon-electronics-dataset/Preprocessed_dataset/

print("Cleaned dataset uploaded to: s3://amazon-electronics-dataset/Preprocessed_dataset/")
