==> notebooks/03_grouping_data.ipynb <==
!pip install polars

import polars as pl
import os
import sys

sys.path.append("../src")
from grouping_data import clean_types, group_reviews_by_product

s3_path = "s3://amazon-electronics-dataset/Preprocessed_dataset/electronics_clean.csv"

df_lazy = pl.scan_csv(
    s3_path,
    schema_overrides={"is_verified": pl.Utf8, "product_id": pl.Utf8},
    infer_schema_length=0,
    ignore_errors=True
)

df_head = df_lazy.collect()
df_head

#Clean dtypes
df=clean_types(df_head)
df.head()

#Grouping data
grouped_data=group_reviews_by_product(df_head)
grouped_data.head()

#Save to S3
local_output = "product_grouped.csv"
s3_output = "s3://amazon-electronics-dataset/grouping_dataset/product_grouped.csv"

grouped_data.write_csv(local_output)

os.system(f"aws s3 cp {local_output} {s3_output}")

print("Saved grouped dataset to:", s3_output)
