==> notebooks/04_abstractive_summarization.ipynb <==
# Abstractive Summarization

from sagemaker.processing import ScriptProcessor, ProcessingInput, ProcessingOutput
from sagemaker import get_execution_role
import sagemaker

role = get_execution_role()

processor = ScriptProcessor(
    image_uri="763104351884.dkr.ecr.us-east-1.amazonaws.com/pytorch-training:2.1.0-gpu-py310",
    command=["python3"],
    instance_type="ml.g5.12xlarge",
    instance_count=1,
    role=role
)

processor.run(
    code="../run_processing.py",
    inputs=[
        ProcessingInput(
            source="s3://amazon-electronics-dataset/grouping_dataset/product_grouped.csv",
            destination="/opt/ml/processing/input"
        ),
        ProcessingInput(
            source="../requirements.txt",
            destination="/opt/ml/processing/input/requirements"
        ),
        ProcessingInput(
            source="../src",
            destination="/opt/ml/processing/src"
        ),
        ProcessingInput(
            source="s3://amazon-electronics-dataset/pegasus_model/",
            destination="/opt/ml/processing/pegasus_model"
        )
    ],
    outputs=[
        ProcessingOutput(
            source="/opt/ml/processing/output",
            destination="s3://amazon-electronics-dataset/summarized_dataset/"
        ),
        ProcessingOutput(
            source="/opt/ml/processing/checkpoints",
            destination="s3://amazon-electronics-dataset/checkpoint_files/"
        )
    ],
    arguments=["--top_k", "30700"]
)

print("Job Started")
